http://bugs.uclibc.org/view.php?id=288
This updates to the current version of libsysfs, and also changes
to the current version of the package.

NOTE:  this package still has a problem, in that the header files
it generates are not made available to other packages.  The library
itself is installed in the target filesystem; but programs like
"brctl" won't know to use it, since the header isn't visible.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>