package/freetype: Convert to Makefile.autotools.in
Use AutoTools.in to build freetype.
Slight issue with freetype installation to TARGET_DIR.
Doing an install results in far too many files being copied in.
So for now I do what the old makefile did which is copy the .so files 
and links manually rather than using a make install.
It also uses the autotools.in additions to not use the generic libtool patch and
use a local one instead.

Daniel Laird