ace_of_penguins: mark as BROKEN
This package is BROKEN in the way it compiles make-imglib and text2c
for the host while being in cross-compiled mode. It does not use
CC_FOR_BUILD and CFLAGS_FOR_BUILD as it should.

No upstream release has been made since 2001 and since it is a
relatively useless package, I don't think it's worth the trouble to
sanitize its build system.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>