libidn: fix build with config.cache, autoconf and emacs
libidn contains a few elisp files, and it's configure script checks for
emacs to know if it should install them. This is not important for BR as
we don't have emacs, but configure fails if it's available on the host,
config.cache is used and autoconf has already been used as that loads
EMACS="no" into the cache.

Fix it by also setting EMACS="no" here.

At the same time, fix up trailing spaces in Config.in.

Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>