Config.in.mirrors: Use sources.buildroot.net as default backup site
The buildroot/busybox/uClibc VM is running low on disk space, and we've
been asked to move the source mirrors off-site.

A redirect has been setup between the old buildroot.net/downloads/sources/
and sources.buildroot.net, so old .configs continue to work, but we might
as well use the official one now.

Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>