microwin: remove deprecated package
Microwin is disabled in Buildroot since January 2007
(726abd4e519f97eda6f6dfc1b6b9505fc37e8b23) and nobody
complained. Moreover, the upstream project hasn't done any release
since five years, so let's remove this package, as announced in the
notes of 2010.08 release.

The project maintainer has announced in June 2010 that he is working
again on the project, and he has switched to Git. When a new release
is done, we can re-introduce it if people need it (this old package
needed serious work to be converted to the new package infrastructure
anyway).

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>