tslib: Enable raw module in ts.conf
By default there is no module activated in ts.conf. According to the
ts.conf previously attached in the tslib package dir (which was never
installed) the default module of choice was the raw module.

This patch enables the raw module in the ts.conf file in the tslib
source tree, which will be installed in the target.

Signed-off-by: Daniel Nyström <daniel.nystrom@timeterminal.se>
Acked-by: Paulius Zaleckas <paulius.zaleckas@gmail.com>
Acked-by: Luca Ceresoli <luca@lucaceresoli.net>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>