autoconf: use dummy autopoint
We currently don't build a host-gettext package, so when a package using
gettext is autoreconf'ed, we end up using autopoint from the host.

Autopoint unfortunately requires CVS, so if that isn't available autoreconf
fails.

From reading:

http://www.gnu.org/software/gnulib/manual/html_node/gettextize-and-autopoint.html

It seems that we don't really need to rerun autopoint, so fix it by
passing autopoint=/bin/true instead.

Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>