util-linux: fix install when locales are enabled
util-linux uses mkinstalldirs to install .po files when locales are
enabled, but the definition of MKINSTALLDIRS comes from a tweaked
nls.m4 (based on the one from gettext).

When we autoreconf the package, we end up using the system (staging)
version of nls.m4, so MKINSTALLDIRS doesn't get defined.

Fix it by passing a definition of MKINSTALLDIRS to make during the
install steps.

Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>