package: fix bootloader building
The changes in e055aea216cff4d12cf55eab2e35ca1459edaa91 and
0849e8193ef429f29945b8c2a54eafc8635e0969 have broken the bootloader
support.

This is because the 'pkgparentdir' macro returns the string 'boot/'
for bootloaders, while the GENTARGETS_INNER macro was testing this
against 'boot' to know if the configuration option to check for this
package is BR2_PACKAGE_* or BR2_TARGET_*.

So, instead of testing against 'boot', we test against 'boot/'.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>