Make BCM DirectFB the real DirectFB target.
This removes the weird indirect target previously required for cairo &
directfb to compile to build the webkit & netflix library.  Instead,
we simply remove the build system for the opensource DirectFB and
replace it with a call to the Broadcom-wrapped DirectFB instance.  The
explicit dependency to DirectFB for bcm_apps is removed, instead
relying on the Broadcom build system's dependency management to have
built the DirectFB target required for the browser and netflix.
However, we do turn on the DirectFB package, so that other buildroot
packages (like Cairo) have their dependencies fulfilled.

Change-Id: Ieef860d25c41e3659026e303ae9f6ae313e9f952