Support for squashfs for installer
Add BR2_PACKAGE_GOOGLE_VMLINUZ, BR2_TARGET_ROOTFS_SQUASHFS_UBI to all bruno configs
Additionally bring bruno_gfhd100b0_defconfig in sync with the A0 version
Build a UBI-wrapped squashfs for ginstall to use
Create vmlinuz in a buildroot package, and wrap it in a UBI header.
Not yet using the vmlinuz.ubi in the installer, the bootloader is not ready yet.

Change-Id: Ib450251860299229cad1dd95845839bece58a8cd