Manual apply upstream patches.
	jnewlin:  Manually patched in these two changes.
	Really just need the second one, but it depended on the previous
	change.

First change:
external-toolchain: add support for recent Linaro toolchains
    Starting from 2012.03, the Linaro toolchains have separated the GCC
    support libraries (libstdc++, libgcc_s) from the sysroot itself. So we
    no longer have the case where all libraries are inside the sysroot, as
    we had for all the previously supported toolchains.

    Therefore, we add some logic to detect if such a separate directory is
    used for GCC support libraries, and if it's the case, we make sure
    that we take into account this directory when creating our own
    sysroot, and when copying libraries to the target filesystem.

    Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>

Second change:
Improve external toolchain logic to support IA32 Sourcery CodeBench toolchain

The IA32 Sourcery CodeBench toolchain has a relatively special
structure, with the following multilib variants:

 * Intel Pentium 4, 32 bits, the multilib variant is in ./ relative to
   the main sysroot, with the libraries in the lib/ directory.

 * Intel Xeon Nocona, 64 bits, the multilib variant is in ./ relative
   to the main sysroot, with the libraries in the lib64/ directory.

 * Intel Atom 32 bits, the multilib variant is in atom/ relative to
   the main sysroot, with the libraries in the lib/ directory.

 * Intel Core 2 64 bits, the multilib variant is in core2/ relative to
   the main sysroot, with the libraries in lib64/ directory.

So the first two variants are in the same sysroot, only the name of
the directory for the libraries is different.

Therefore, we introduce a new ARCH_LIB_DIR variable, which contains
either 'lib' or 'lib64'. This variable is defined according to the
location of the libc.a file for the selected multilib variant, and is
then used when copying the libraries to the target and to the staging
directory.

In addition to this, we no longer use the -print-multi-directory to
get the ARCH_SUBDIR, since in the case of the 64 bits variants of this
toolchain, it returns just '64' and not a real path. Instead, we
simply compute the difference between the arch-specific sysroot and
the main sysroot.

We also take that opportunity to expand the documentation on the
meaning of the different variables.

Change-Id: Ia5e49a5ae238adf0b40e32e56d481f4954424e79
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>