builder.py: check LOAS certificate whether or not -x is given.
Technically it's not needed in order to make a platform-only build, but
that's just confusing (particularly if you later run 'builder.py -b' and
wonder why you need LOAS now, because you forgot to specify -x, because -x
has no other effect on -b but to disable this check).

In other words it's an extra code path we have to keep maintained.  Let's
keep it simple.

Change-Id: I137b1bf4ec39fea85d8119e378728b1f404e8bd8