util-linux/logger.c: default priority must include LOG_USER.
glibc doesn't let you log with LOG_KERN, but uClibc does not seem to have
that restriction.  To make sure logs that don't specify a facility or
priority don't accidentally end up as LOG_KERN, add LOG_USER to the
default.

(This was already happening if you *do* specify the -p option.)

Change-Id: I8aea1ba0bd10cfa8ece20a384d1c66f11d4436d0