S50dropbear: try to persist the server cert.
S50dropbear auto-generates a certificate if the server doesn't already have
one, but it would always dump it in /tmp/dropbear, which means we got a
different certificate on every boot.  Not only did this add a couple of
seconds to the boot process, it also means you have to remove the server's
cert from your ~/.ssh/known_hosts all the time.

Instead, if /config is mounted read-write, save the certs there.  We still
fall back to /tmp just in case /config isn't available, since if that
happens, you might *really* wish there was an sshd to use to debug why.

Change-Id: I24387d3b15e6dcd9629578742eace7940af20bd6