simpleramfs: allow login=1 *or* debug=1 to enable console logins.
Unfortunately, debug=1 is also interpreted by the *kernel* if set, and it
starts spewing even debug-level kernel messages to the console.  That might
be useful sometimes, but not if all you want is a login prompt.  So in
simpleramfs, look for a login= kernel parameter, and if set, enable
console/sshd logins.  For backward compatiblity, debug= keeps working too
(and still spews lots of extra messages).

Change-Id: I90f426a5da780fc71321aa655aa529a95f643263